+++ src/components.d.ts
@@ -5,9 +5,10 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
+import { IAccountScreenData } from "./types/address-select.types";
 import { IconDefinition } from "@fortawesome/free-solid-svg-icons";
 import { CustomToastType, IComponentToast, ISimpleToast } from "./components/functional/toasts-list/components/transaction-toast/transaction-toast.type";
-import { IAccountScreenData, IConfirmScreenData, IConnectScreenData, ILedgerConnectPanelData } from "./components/functional/ledger-connect/ledger-connect.types";
+import { IConfirmScreenData, IConnectScreenData, ILedgerConnectPanelData } from "./components/functional/ledger-connect/ledger-connect.types";
 import { IEventBus } from "./utils/EventBus";
 import { IProviderBase, ProviderTypeEnum } from "./types/provider.types";
 import { DecodeMethodEnum } from "./components/functional/sign-transactions-panel/sign-transactions-panel.types";
@@ -19,9 +20,10 @@ import { LocalJSX as JSX } from "@stencil/core";
 import { IToastDataState, ITransactionProgressState } from "./components/functional/toasts-list/components/transaction-toast/transaction-toast.type";
 import { TransactionValueType } from "./components/controlled/transactions-table/transactions-table.type";
 import { IEventBus as IEventBus1, unknown as IWalletConnectPanelData } from "./components.d";
+export { IAccountScreenData } from "./types/address-select.types";
 export { IconDefinition } from "@fortawesome/free-solid-svg-icons";
 export { CustomToastType, IComponentToast, ISimpleToast } from "./components/functional/toasts-list/components/transaction-toast/transaction-toast.type";
-export { IAccountScreenData, IConfirmScreenData, IConnectScreenData, ILedgerConnectPanelData } from "./components/functional/ledger-connect/ledger-connect.types";
+export { IConfirmScreenData, IConnectScreenData, ILedgerConnectPanelData } from "./components/functional/ledger-connect/ledger-connect.types";
 export { IEventBus } from "./utils/EventBus";
 export { IProviderBase, ProviderTypeEnum } from "./types/provider.types";
 export { DecodeMethodEnum } from "./components/functional/sign-transactions-panel/sign-transactions-panel.types";
@@ -34,6 +36,10 @@ export { IToastDataState, ITransactionProgressState } from "./components/functio
 export { TransactionValueType } from "./components/controlled/transactions-table/transactions-table.type";
 export { IEventBus as IEventBus1, unknown as IWalletConnectPanelData } from "./components.d";
 export namespace Components {
+    interface MvxAddressSelect {
+        "accountScreenData": IAccountScreenData;
+        "selectedIndex": number;
+    }
     interface MvxAnglesLeftIcon {
     }
     interface MvxAnglesRightIcon {
@@ -58,14 +64,8 @@ export namespace Components {
     interface MvxCloseIcon {
     }
     interface MvxCopyButton {
-        /**
-          * @default 'copy-button'
-         */
         "class"?: string;
         "copyIcon"?: IconDefinition | string;
-        /**
-          * @default 'copy-button-icon'
-         */
         "iconClass"?: string;
         "successIcon"?: IconDefinition | string;
         "text": string;
@@ -91,9 +91,6 @@ export namespace Components {
         "width"?: number;
     }
     interface MvxFaIcon {
-        /**
-          * @default 'fa-icon'
-         */
         "class"?: string;
         "description"?: string;
         "icon": IconDefinition | string;
@@ -105,9 +102,6 @@ export namespace Components {
         "isValid": boolean;
         "label"?: string;
         "labelClass"?: string;
-        /**
-          * @default true
-         */
         "showLabel"?: boolean;
         "valueDecimal": string;
         "valueInteger": string;
@@ -115,17 +109,10 @@ export namespace Components {
     interface MvxGenericToast {
         "toast": CustomToastType;
     }
-    interface MvxLedgerAddresses {
-        "accountScreenData": IAccountScreenData;
-        "selectedIndex": number;
-    }
     interface MvxLedgerConfirm {
         "confirmScreenData": IConfirmScreenData;
     }
     interface MvxLedgerConnect {
-        /**
-          * @default {     accountScreenData: null,     confirmScreenData: null,     connectScreenData: {},   }
-         */
         "data": ILedgerConnectPanelData;
         "getEventBus": () => Promise<IEventBus>;
     }
@@ -134,9 +121,6 @@ export namespace Components {
     }
     interface MvxLedgerIntro {
         "connectScreenData"?: IConnectScreenData;
-        /**
-          * @default false
-         */
         "isAwaiting"?: boolean;
     }
     interface MvxLedgerProviderIcon {
@@ -157,26 +141,14 @@ export namespace Components {
     }
     interface MvxPagination {
         "class"?: string;
-        /**
-          * @default 1
-         */
         "currentPage": number;
-        /**
-          * @default false
-         */
         "isDisabled"?: boolean;
         "totalPages": number;
     }
     interface MvxPaginationEllipsis {
-        /**
-          * @default false
-         */
         "isActive": boolean;
     }
     interface MvxPaginationEllipsisForm {
-        /**
-          * @default false
-         */
         "isVisible": boolean;
         "maxPageToSearchFor": number;
     }
@@ -194,43 +166,28 @@ export namespace Components {
         "class"?: string;
     }
     interface MvxProviderIdleScreen {
-        /**
-          * @default ''
-         */
         "introText": string;
-        /**
-          * @default 'Requesting Connection'
-         */
         "introTitle": string;
-        /**
-          * @default null
-         */
         "provider": IProviderBase | null;
     }
     interface MvxSidePanel {
         "hasBackButton"?: boolean;
-        /**
-          * @default false
-         */
         "isOpen": boolean;
         "panelClassName"?: string;
         "panelTitle": string;
-        /**
-          * @default true
-         */
         "showHeader"?: boolean;
     }
     interface MvxSidePanelHeader {
-        /**
-          * @default true
-         */
         "hasLeftButton"?: boolean;
-        /**
-          * @default true
-         */
         "hasRightButton"?: boolean;
         "panelClassName"?: string;
         "panelTitle": string;
+    }
+    interface MvxSidePanelSwiper {
+        "close": () => Promise<void>;
+        "open": boolean;
+        "openToSnapPoint": (snapIndex?: number) => Promise<void>;
+        "sidePanelIdentifier": string;
     }
     interface MvxSignTransactionsAdvanced {
         "data": string;
@@ -241,13 +198,7 @@ export namespace Components {
         "highlight"?: string;
     }
     interface MvxSignTransactionsAdvancedDataDecode {
-        /**
-          * @default DecodeMethodEnum.decimal
-         */
         "currentDecodeMethod": DecodeMethodEnum;
-        /**
-          * @default false
-         */
         "isToggled": boolean;
     }
     interface MvxSignTransactionsFooter {
@@ -260,13 +211,7 @@ export namespace Components {
         "identifier": string;
         "interactor": string;
         "interactorIconUrl": string;
-        /**
-          * @default false
-         */
         "isApp": boolean;
-        /**
-          * @default '~$0.00078'
-         */
         "networkFee": string;
         "tokenIconUrl": string;
         "usdValue": string;
@@ -295,14 +240,8 @@ export namespace Components {
     }
     interface MvxTooltip {
         "class"?: string;
-        /**
-          * @default 'top'
-         */
         "position": 'top' | 'bottom';
         "trigger": HTMLElement;
-        /**
-          * @default false
-         */
         "triggerOnClick"?: boolean;
     }
     interface MvxTransactionAccount {
@@ -310,9 +249,6 @@ export namespace Components {
         "class"?: string;
         "dataTestId"?: string;
         "scope": 'receiver' | 'sender';
-        /**
-          * @default false
-         */
         "showLockedAccounts": boolean;
     }
     interface MvxTransactionAccountName {
@@ -357,19 +293,10 @@ export namespace Components {
     }
     interface MvxTransactionToast {
         "fullWidth"?: boolean;
-        /**
-          * @default ''
-         */
         "processedTransactionsStatus": string | JSX.Element;
         "toastDataState": IToastDataState;
-        /**
-          * @default ''
-         */
         "toastId": string;
         "transactionProgressState"?: ITransactionProgressState;
-        /**
-          * @default []
-         */
         "transactions": ITransactionListItem[];
         "wrapperClass": string;
     }
@@ -380,9 +307,6 @@ export namespace Components {
         "transactions": ITransactionListItem[];
     }
     interface MvxTransactionToastDetails {
-        /**
-          * @default 5
-         */
         "maxShownTransactions": number;
         "processedTransactionsStatus"?: string | JSX.Element;
         "transactionClass": string;
@@ -393,9 +317,6 @@ export namespace Components {
         "index": string;
         "link": string;
         "status"?: string;
-        /**
-          * @default 'transaction-details-list-item'
-         */
         "transactionClass"?: string;
     }
     interface MvxTransactionToastProgress {
@@ -419,6 +340,7 @@ export namespace Components {
     }
     interface MvxUnlockButton {
         "class"?: string;
+        "dataTestId"?: string;
         "icon"?: HTMLElement;
         "iconUrl": string;
         "label": string;
@@ -433,9 +355,6 @@ export namespace Components {
     }
     interface MvxUnlockPanelGroup {
         "class"?: string;
-        /**
-          * @default []
-         */
         "providers": IProviderBase[];
     }
     interface MvxUnlockProviderButton {
@@ -443,14 +362,8 @@ export namespace Components {
         "provider": IProviderBase<ProviderTypeEnum>;
     }
     interface MvxWalletConnect {
-        /**
-          * @default { wcURI: '' }
-         */
         "data": IWalletConnectPanelData;
         "getEventBus": () => Promise<IEventBus>;
-        /**
-          * @default ''
-         */
         "qrCodeSvg": string;
     }
     interface MvxWalletConnectAppGalleryIcon {
@@ -467,13 +380,7 @@ export namespace Components {
     }
     interface MvxWalletConnectScan {
         "class"?: string;
-        /**
-          * @default ''
-         */
         "qrCodeSvg": string;
-        /**
-          * @default ''
-         */
         "walletConnectDeepLink"?: string;
     }
     interface MvxWalletProviderIcon {
@@ -482,14 +389,15 @@ export namespace Components {
     interface MvxXaliasProviderIcon {
     }
     interface MvxXportalDownloadQrIcon {
-        /**
-          * @default ''
-         */
         "class"?: string;
     }
     interface MvxXportalQrCodePreloader {
         "class"?: string;
     }
+}
+export interface MvxAddressSelectCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLMvxAddressSelectElement;
 }
 export interface MvxCustomToastCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -498,10 +406,6 @@ export interface MvxCustomToastCustomEvent<T> extends CustomEvent<T> {
 export interface MvxGenericToastCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLMvxGenericToastElement;
-}
-export interface MvxLedgerAddressesCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLMvxLedgerAddressesElement;
 }
 export interface MvxLedgerIntroCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -526,6 +430,10 @@ export interface MvxSidePanelCustomEvent<T> extends CustomEvent<T> {
 export interface MvxSidePanelHeaderCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLMvxSidePanelHeaderElement;
+}
+export interface MvxSidePanelSwiperCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLMvxSidePanelSwiperElement;
 }
 export interface MvxSimpleToastCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -552,6 +460,25 @@ export interface MvxWalletConnectScanCustomEvent<T> extends CustomEvent<T> {
     target: HTMLMvxWalletConnectScanElement;
 }
 declare global {
+    interface HTMLMvxAddressSelectElementEventMap {
+        "accessWallet": any;
+        "selectAccount": any;
+        "pageChange": number;
+    }
+    interface HTMLMvxAddressSelectElement extends Components.MvxAddressSelect, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLMvxAddressSelectElementEventMap>(type: K, listener: (this: HTMLMvxAddressSelectElement, ev: MvxAddressSelectCustomEvent<HTMLMvxAddressSelectElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLMvxAddressSelectElementEventMap>(type: K, listener: (this: HTMLMvxAddressSelectElement, ev: MvxAddressSelectCustomEvent<HTMLMvxAddressSelectElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLMvxAddressSelectElement: {
+        prototype: HTMLMvxAddressSelectElement;
+        new (): HTMLMvxAddressSelectElement;
+    };
     interface HTMLMvxAnglesLeftIconElement extends Components.MvxAnglesLeftIcon, HTMLStencilElement {
     }
     var HTMLMvxAnglesLeftIconElement: {
@@ -681,25 +608,6 @@ declare global {
     var HTMLMvxGenericToastElement: {
         prototype: HTMLMvxGenericToastElement;
         new (): HTMLMvxGenericToastElement;
-    };
-    interface HTMLMvxLedgerAddressesElementEventMap {
-        "accessWallet": any;
-        "selectAccount": any;
-        "pageChange": number;
-    }
-    interface HTMLMvxLedgerAddressesElement extends Components.MvxLedgerAddresses, HTMLStencilElement {
-        addEventListener<K extends keyof HTMLMvxLedgerAddressesElementEventMap>(type: K, listener: (this: HTMLMvxLedgerAddressesElement, ev: MvxLedgerAddressesCustomEvent<HTMLMvxLedgerAddressesElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLMvxLedgerAddressesElementEventMap>(type: K, listener: (this: HTMLMvxLedgerAddressesElement, ev: MvxLedgerAddressesCustomEvent<HTMLMvxLedgerAddressesElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
-    }
-    var HTMLMvxLedgerAddressesElement: {
-        prototype: HTMLMvxLedgerAddressesElement;
-        new (): HTMLMvxLedgerAddressesElement;
     };
     interface HTMLMvxLedgerConfirmElement extends Components.MvxLedgerConfirm, HTMLStencilElement {
     }
@@ -883,6 +791,24 @@ declare global {
     var HTMLMvxSidePanelHeaderElement: {
         prototype: HTMLMvxSidePanelHeaderElement;
         new (): HTMLMvxSidePanelHeaderElement;
+    };
+    interface HTMLMvxSidePanelSwiperElementEventMap {
+        "sheetDismiss": void;
+        "sheetSnapChange": { index: number; snapPoint: string };
+    }
+    interface HTMLMvxSidePanelSwiperElement extends Components.MvxSidePanelSwiper, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLMvxSidePanelSwiperElementEventMap>(type: K, listener: (this: HTMLMvxSidePanelSwiperElement, ev: MvxSidePanelSwiperCustomEvent<HTMLMvxSidePanelSwiperElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLMvxSidePanelSwiperElementEventMap>(type: K, listener: (this: HTMLMvxSidePanelSwiperElement, ev: MvxSidePanelSwiperCustomEvent<HTMLMvxSidePanelSwiperElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLMvxSidePanelSwiperElement: {
+        prototype: HTMLMvxSidePanelSwiperElement;
+        new (): HTMLMvxSidePanelSwiperElement;
     };
     interface HTMLMvxSignTransactionsAdvancedElement extends Components.MvxSignTransactionsAdvanced, HTMLStencilElement {
     }
@@ -1239,6 +1165,7 @@ declare global {
         new (): HTMLMvxXportalQrCodePreloaderElement;
     };
     interface HTMLElementTagNameMap {
+        "mvx-address-select": HTMLMvxAddressSelectElement;
         "mvx-angles-left-icon": HTMLMvxAnglesLeftIconElement;
         "mvx-angles-right-icon": HTMLMvxAnglesRightIconElement;
         "mvx-arrow-right-icon": HTMLMvxArrowRightIconElement;
@@ -1257,7 +1184,6 @@ declare global {
         "mvx-fa-icon": HTMLMvxFaIconElement;
         "mvx-format-amount": HTMLMvxFormatAmountElement;
         "mvx-generic-toast": HTMLMvxGenericToastElement;
-        "mvx-ledger-addresses": HTMLMvxLedgerAddressesElement;
         "mvx-ledger-confirm": HTMLMvxLedgerConfirmElement;
         "mvx-ledger-connect": HTMLMvxLedgerConnectElement;
         "mvx-ledger-icon": HTMLMvxLedgerIconElement;
@@ -1277,6 +1203,7 @@ declare global {
         "mvx-provider-idle-screen": HTMLMvxProviderIdleScreenElement;
         "mvx-side-panel": HTMLMvxSidePanelElement;
         "mvx-side-panel-header": HTMLMvxSidePanelHeaderElement;
+        "mvx-side-panel-swiper": HTMLMvxSidePanelSwiperElement;
         "mvx-sign-transactions-advanced": HTMLMvxSignTransactionsAdvancedElement;
         "mvx-sign-transactions-advanced-data": HTMLMvxSignTransactionsAdvancedDataElement;
         "mvx-sign-transactions-advanced-data-decode": HTMLMvxSignTransactionsAdvancedDataDecodeElement;
@@ -1328,6 +1255,13 @@ declare global {
     }
 }
 declare namespace LocalJSX {
+    interface MvxAddressSelect {
+        "accountScreenData"?: IAccountScreenData;
+        "onAccessWallet"?: (event: MvxAddressSelectCustomEvent<any>) => void;
+        "onPageChange"?: (event: MvxAddressSelectCustomEvent<number>) => void;
+        "onSelectAccount"?: (event: MvxAddressSelectCustomEvent<any>) => void;
+        "selectedIndex"?: number;
+    }
     interface MvxAnglesLeftIcon {
     }
     interface MvxAnglesRightIcon {
@@ -1352,14 +1286,8 @@ declare namespace LocalJSX {
     interface MvxCloseIcon {
     }
     interface MvxCopyButton {
-        /**
-          * @default 'copy-button'
-         */
         "class"?: string;
         "copyIcon"?: IconDefinition | string;
-        /**
-          * @default 'copy-button-icon'
-         */
         "iconClass"?: string;
         "successIcon"?: IconDefinition | string;
         "text"?: string;
@@ -1386,9 +1314,6 @@ declare namespace LocalJSX {
         "width"?: number;
     }
     interface MvxFaIcon {
-        /**
-          * @default 'fa-icon'
-         */
         "class"?: string;
         "description"?: string;
         "icon"?: IconDefinition | string;
@@ -1400,9 +1325,6 @@ declare namespace LocalJSX {
         "isValid"?: boolean;
         "label"?: string;
         "labelClass"?: string;
-        /**
-          * @default true
-         */
         "showLabel"?: boolean;
         "valueDecimal"?: string;
         "valueInteger"?: string;
@@ -1411,20 +1333,10 @@ declare namespace LocalJSX {
         "onDeleteToast"?: (event: MvxGenericToastCustomEvent<string>) => void;
         "toast"?: CustomToastType;
     }
-    interface MvxLedgerAddresses {
-        "accountScreenData"?: IAccountScreenData;
-        "onAccessWallet"?: (event: MvxLedgerAddressesCustomEvent<any>) => void;
-        "onPageChange"?: (event: MvxLedgerAddressesCustomEvent<number>) => void;
-        "onSelectAccount"?: (event: MvxLedgerAddressesCustomEvent<any>) => void;
-        "selectedIndex"?: number;
-    }
     interface MvxLedgerConfirm {
         "confirmScreenData"?: IConfirmScreenData;
     }
     interface MvxLedgerConnect {
-        /**
-          * @default {     accountScreenData: null,     confirmScreenData: null,     connectScreenData: {},   }
-         */
         "data"?: ILedgerConnectPanelData;
     }
     interface MvxLedgerIcon {
@@ -1432,9 +1344,6 @@ declare namespace LocalJSX {
     }
     interface MvxLedgerIntro {
         "connectScreenData"?: IConnectScreenData;
-        /**
-          * @default false
-         */
         "isAwaiting"?: boolean;
         "onConnect"?: (event: MvxLedgerIntroCustomEvent<any>) => void;
     }
@@ -1454,27 +1363,15 @@ declare namespace LocalJSX {
     }
     interface MvxPagination {
         "class"?: string;
-        /**
-          * @default 1
-         */
         "currentPage"?: number;
-        /**
-          * @default false
-         */
         "isDisabled"?: boolean;
         "onPageChange"?: (event: MvxPaginationCustomEvent<number>) => void;
         "totalPages"?: number;
     }
     interface MvxPaginationEllipsis {
-        /**
-          * @default false
-         */
         "isActive"?: boolean;
     }
     interface MvxPaginationEllipsisForm {
-        /**
-          * @default false
-         */
         "isVisible"?: boolean;
         "maxPageToSearchFor"?: number;
         "onSearch"?: (event: MvxPaginationEllipsisFormCustomEvent<number>) => void;
@@ -1491,49 +1388,34 @@ declare namespace LocalJSX {
         "class"?: string;
     }
     interface MvxProviderIdleScreen {
-        /**
-          * @default ''
-         */
         "introText"?: string;
-        /**
-          * @default 'Requesting Connection'
-         */
         "introTitle"?: string;
         "onAccess"?: (event: MvxProviderIdleScreenCustomEvent<any>) => void;
         "onClose"?: (event: MvxProviderIdleScreenCustomEvent<any>) => void;
-        /**
-          * @default null
-         */
         "provider"?: IProviderBase | null;
     }
     interface MvxSidePanel {
         "hasBackButton"?: boolean;
-        /**
-          * @default false
-         */
         "isOpen"?: boolean;
         "onBack"?: (event: MvxSidePanelCustomEvent<void>) => void;
         "onClose"?: (event: MvxSidePanelCustomEvent<void>) => void;
         "panelClassName"?: string;
         "panelTitle"?: string;
-        /**
-          * @default true
-         */
         "showHeader"?: boolean;
     }
     interface MvxSidePanelHeader {
-        /**
-          * @default true
-         */
         "hasLeftButton"?: boolean;
-        /**
-          * @default true
-         */
         "hasRightButton"?: boolean;
         "onLeftButtonClick"?: (event: MvxSidePanelHeaderCustomEvent<any>) => void;
         "onRightButtonClick"?: (event: MvxSidePanelHeaderCustomEvent<any>) => void;
         "panelClassName"?: string;
         "panelTitle"?: string;
+    }
+    interface MvxSidePanelSwiper {
+        "onSheetDismiss"?: (event: MvxSidePanelSwiperCustomEvent<void>) => void;
+        "onSheetSnapChange"?: (event: MvxSidePanelSwiperCustomEvent<{ index: number; snapPoint: string }>) => void;
+        "open"?: boolean;
+        "sidePanelIdentifier"?: string;
     }
     interface MvxSignTransactionsAdvanced {
         "data"?: string;
@@ -1544,13 +1426,7 @@ declare namespace LocalJSX {
         "highlight"?: string;
     }
     interface MvxSignTransactionsAdvancedDataDecode {
-        /**
-          * @default DecodeMethodEnum.decimal
-         */
         "currentDecodeMethod"?: DecodeMethodEnum;
-        /**
-          * @default false
-         */
         "isToggled"?: boolean;
     }
     interface MvxSignTransactionsFooter {
@@ -1563,13 +1439,7 @@ declare namespace LocalJSX {
         "identifier"?: string;
         "interactor"?: string;
         "interactorIconUrl"?: string;
-        /**
-          * @default false
-         */
         "isApp"?: boolean;
-        /**
-          * @default '~$0.00078'
-         */
         "networkFee"?: string;
         "tokenIconUrl"?: string;
         "usdValue"?: string;
@@ -1597,14 +1467,8 @@ declare namespace LocalJSX {
     interface MvxTooltip {
         "class"?: string;
         "onTriggerRender"?: (event: MvxTooltipCustomEvent<boolean>) => void;
-        /**
-          * @default 'top'
-         */
         "position"?: 'top' | 'bottom';
         "trigger"?: HTMLElement;
-        /**
-          * @default false
-         */
         "triggerOnClick"?: boolean;
     }
     interface MvxTransactionAccount {
@@ -1612,9 +1476,6 @@ declare namespace LocalJSX {
         "class"?: string;
         "dataTestId"?: string;
         "scope"?: 'receiver' | 'sender';
-        /**
-          * @default false
-         */
         "showLockedAccounts"?: boolean;
     }
     interface MvxTransactionAccountName {
@@ -1660,19 +1521,10 @@ declare namespace LocalJSX {
     interface MvxTransactionToast {
         "fullWidth"?: boolean;
         "onDeleteToast"?: (event: MvxTransactionToastCustomEvent<void>) => void;
-        /**
-          * @default ''
-         */
         "processedTransactionsStatus"?: string | JSX.Element;
         "toastDataState"?: IToastDataState;
-        /**
-          * @default ''
-         */
         "toastId"?: string;
         "transactionProgressState"?: ITransactionProgressState;
-        /**
-          * @default []
-         */
         "transactions"?: ITransactionListItem[];
         "wrapperClass"?: string;
     }
@@ -1684,9 +1536,6 @@ declare namespace LocalJSX {
         "transactions"?: ITransactionListItem[];
     }
     interface MvxTransactionToastDetails {
-        /**
-          * @default 5
-         */
         "maxShownTransactions"?: number;
         "processedTransactionsStatus"?: string | JSX.Element;
         "transactionClass"?: string;
@@ -1697,9 +1546,6 @@ declare namespace LocalJSX {
         "index"?: string;
         "link"?: string;
         "status"?: string;
-        /**
-          * @default 'transaction-details-list-item'
-         */
         "transactionClass"?: string;
     }
     interface MvxTransactionToastProgress {
@@ -1723,6 +1569,7 @@ declare namespace LocalJSX {
     }
     interface MvxUnlockButton {
         "class"?: string;
+        "dataTestId"?: string;
         "icon"?: HTMLElement;
         "iconUrl"?: string;
         "label"?: string;
@@ -1736,9 +1583,6 @@ declare namespace LocalJSX {
     interface MvxUnlockPanelGroup {
         "class"?: string;
         "onLogin"?: (event: MvxUnlockPanelGroupCustomEvent<IProviderBase>) => void;
-        /**
-          * @default []
-         */
         "providers"?: IProviderBase[];
     }
     interface MvxUnlockProviderButton {
@@ -1746,13 +1590,7 @@ declare namespace LocalJSX {
         "provider"?: IProviderBase<ProviderTypeEnum>;
     }
     interface MvxWalletConnect {
-        /**
-          * @default { wcURI: '' }
-         */
         "data"?: IWalletConnectPanelData;
-        /**
-          * @default ''
-         */
         "qrCodeSvg"?: string;
     }
     interface MvxWalletConnectAppGalleryIcon {
@@ -1770,13 +1608,7 @@ declare namespace LocalJSX {
     interface MvxWalletConnectScan {
         "class"?: string;
         "onDownloadClick"?: (event: MvxWalletConnectScanCustomEvent<any>) => void;
-        /**
-          * @default ''
-         */
         "qrCodeSvg"?: string;
-        /**
-          * @default ''
-         */
         "walletConnectDeepLink"?: string;
     }
     interface MvxWalletProviderIcon {
@@ -1785,15 +1617,13 @@ declare namespace LocalJSX {
     interface MvxXaliasProviderIcon {
     }
     interface MvxXportalDownloadQrIcon {
-        /**
-          * @default ''
-         */
         "class"?: string;
     }
     interface MvxXportalQrCodePreloader {
         "class"?: string;
     }
     interface IntrinsicElements {
+        "mvx-address-select": MvxAddressSelect;
         "mvx-angles-left-icon": MvxAnglesLeftIcon;
         "mvx-angles-right-icon": MvxAnglesRightIcon;
         "mvx-arrow-right-icon": MvxArrowRightIcon;
@@ -1812,7 +1642,6 @@ declare namespace LocalJSX {
         "mvx-fa-icon": MvxFaIcon;
         "mvx-format-amount": MvxFormatAmount;
         "mvx-generic-toast": MvxGenericToast;
-        "mvx-ledger-addresses": MvxLedgerAddresses;
         "mvx-ledger-confirm": MvxLedgerConfirm;
         "mvx-ledger-connect": MvxLedgerConnect;
         "mvx-ledger-icon": MvxLedgerIcon;
@@ -1832,6 +1661,7 @@ declare namespace LocalJSX {
         "mvx-provider-idle-screen": MvxProviderIdleScreen;
         "mvx-side-panel": MvxSidePanel;
         "mvx-side-panel-header": MvxSidePanelHeader;
+        "mvx-side-panel-swiper": MvxSidePanelSwiper;
         "mvx-sign-transactions-advanced": MvxSignTransactionsAdvanced;
         "mvx-sign-transactions-advanced-data": MvxSignTransactionsAdvancedData;
         "mvx-sign-transactions-advanced-data-decode": MvxSignTransactionsAdvancedDataDecode;
@@ -1886,6 +1716,7 @@ export { LocalJSX as JSX };
 declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
+            "mvx-address-select": LocalJSX.MvxAddressSelect & JSXBase.HTMLAttributes<HTMLMvxAddressSelectElement>;
             "mvx-angles-left-icon": LocalJSX.MvxAnglesLeftIcon & JSXBase.HTMLAttributes<HTMLMvxAnglesLeftIconElement>;
             "mvx-angles-right-icon": LocalJSX.MvxAnglesRightIcon & JSXBase.HTMLAttributes<HTMLMvxAnglesRightIconElement>;
             "mvx-arrow-right-icon": LocalJSX.MvxArrowRightIcon & JSXBase.HTMLAttributes<HTMLMvxArrowRightIconElement>;
@@ -1904,7 +1735,6 @@ declare module "@stencil/core" {
             "mvx-fa-icon": LocalJSX.MvxFaIcon & JSXBase.HTMLAttributes<HTMLMvxFaIconElement>;
             "mvx-format-amount": LocalJSX.MvxFormatAmount & JSXBase.HTMLAttributes<HTMLMvxFormatAmountElement>;
             "mvx-generic-toast": LocalJSX.MvxGenericToast & JSXBase.HTMLAttributes<HTMLMvxGenericToastElement>;
-            "mvx-ledger-addresses": LocalJSX.MvxLedgerAddresses & JSXBase.HTMLAttributes<HTMLMvxLedgerAddressesElement>;
             "mvx-ledger-confirm": LocalJSX.MvxLedgerConfirm & JSXBase.HTMLAttributes<HTMLMvxLedgerConfirmElement>;
             "mvx-ledger-connect": LocalJSX.MvxLedgerConnect & JSXBase.HTMLAttributes<HTMLMvxLedgerConnectElement>;
             "mvx-ledger-icon": LocalJSX.MvxLedgerIcon & JSXBase.HTMLAttributes<HTMLMvxLedgerIconElement>;
@@ -1924,6 +1754,7 @@ declare module "@stencil/core" {
             "mvx-provider-idle-screen": LocalJSX.MvxProviderIdleScreen & JSXBase.HTMLAttributes<HTMLMvxProviderIdleScreenElement>;
             "mvx-side-panel": LocalJSX.MvxSidePanel & JSXBase.HTMLAttributes<HTMLMvxSidePanelElement>;
             "mvx-side-panel-header": LocalJSX.MvxSidePanelHeader & JSXBase.HTMLAttributes<HTMLMvxSidePanelHeaderElement>;
+            "mvx-side-panel-swiper": LocalJSX.MvxSidePanelSwiper & JSXBase.HTMLAttributes<HTMLMvxSidePanelSwiperElement>;
             "mvx-sign-transactions-advanced": LocalJSX.MvxSignTransactionsAdvanced & JSXBase.HTMLAttributes<HTMLMvxSignTransactionsAdvancedElement>;
             "mvx-sign-transactions-advanced-data": LocalJSX.MvxSignTransactionsAdvancedData & JSXBase.HTMLAttributes<HTMLMvxSignTransactionsAdvancedDataElement>;
             "mvx-sign-transactions-advanced-data-decode": LocalJSX.MvxSignTransactionsAdvancedDataDecode & JSXBase.HTMLAttributes<HTMLMvxSignTransactionsAdvancedDataDecodeElement>;