+++ src/components/functional/ledger-connect/ledger-connect.types.ts
@@ -1,14 +1,9 @@
+import type { IAccountScreenData } from 'types/address-select.types';
+
 export interface IConnectScreenData {
   customContentMarkup?: string;
   disabled?: boolean;
   error?: string;
-}
-
-export interface IAccountScreenData {
-  accounts: ILedgerAccount[];
-  startIndex: number;
-  addressesPerPage: number;
-  isLoading: boolean;
 }
 
 export interface IConfirmScreenData {
@@ -23,13 +18,6 @@ export interface ILedgerConnectPanelData {
   connectScreenData: IConnectScreenData | null;
   accountScreenData: IAccountScreenData | null;
   confirmScreenData: IConfirmScreenData | null;
-}
-
-export interface ILedgerAccount {
-  address: string;
-  balance: string;
-  usdValue?: string;
-  index: number;
 }
 
 export enum LedgerConnectEventsEnum {