+++ src/components/visual/index.ts
@@ -1,0 +1,1 @@
+export * from './address-select/address-select';