+++ src/types/address-select.types.ts
@@ -1,0 +1,13 @@
+export interface IAccount {
+  address: string;
+  balance: string;
+  usdValue?: string;
+  index: number;
+}
+
+export interface IAccountScreenData {
+  accounts: IAccount[];
+  startIndex: number;
+  addressesPerPage: number;
+  isLoading: boolean;
+}